refactor model imports and user model definition; enhance password validation and hashing
